Update font awesome resource location
At some point the xstatic resources were cleaned up[1].
This broke the previous font-awesome import because the path
changed to reflect the package name, which is font_awesome
(underscore, not hyphen). This fix is in the horizon
chameleoncloud/stable/ocata branch, which we use in prod,
hence the master commit.

[1]: https://github.com/ChameleonCloud/horizon/commit/3f35caf180e1f9fa7e490bf341bf112014d2e6fc