Reference images using absolute paths
Not sure how these relative paths ever worked. It's not supported to
serve images relative to the main CSS directory (unless the images are
explicitly annotated with `{% themable_asset %}` in the template.
Instead, we piggy-back on the existing $static-url variable and add the
theme suffix. This actually requires that our theme be called
'chameleoncloud' when it is configured via `local_settings`, which isn't
ideal, but I can't think of a better way to do this.